+++ CHANGES.rst
@@ -14,6 +14,13 @@ Feature:
 - track ``_elapsed`` time for each executed command ``--junit-xml`` report
   if ``system-out`` is enabled
 
+- track custom properties in ``--junit-xml`` report. For example you can track
+  the time of the time occurred between the end of the previous action and
+  the completion of the following. Basically you can track under the key
+  `load_login` key the time occurred between the click on the submit button
+  and the end of the current command (e.g., click on the menu or text input
+  being able to receive text) using a machine interpretable format
+
 Documentation:
 
 - minor documentation changes